Refactor callbacks. Introduce InlineKeyboardCallbackHandler interface
