+++ tg-bot/src/main/java/polis/keyboards/callbacks/objects/Callback.java
@@ -1,7 +1,5 @@
 package polis.keyboards.callbacks.objects;
 
-import polis.keyboards.callbacks.objects.*;
-
 import java.io.Serializable;
 
 /**
@@ -9,12 +7,5 @@ import java.io.Serializable;
  * @see org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardButton#callbackData
  */
 public abstract sealed class Callback implements Serializable
-        permits
-        AccountCallback,
-        AutopostingCallback,
-        GoBackCallback,
-        GroupCallback,
-        NotificationsCallback,
-        TgChannelCallback,
-        YesNoCallback {
+        permits AccountCallback, AutopostingCallback, GoBackCallback, GroupCallback, NotificationsCallback, ReplyKeyboardCallback, TgChannelCallback, YesNoCallback {
 }
