+++ tg-bot/src/main/java/polis/keyboards/callbacks/objects/ReplyKeyboardCallback.java
@@ -1,0 +1,12 @@
+package polis.keyboards.callbacks.objects;
+
+/**
+ * Message sent using reply keyboard
+ */
+public final class ReplyKeyboardCallback extends Callback {
+    public final String text;
+
+    public ReplyKeyboardCallback(String text) {
+        this.text = text;
+    }
+}
